+++ src/guestbooks/domain/repositories/IGuestbooksRepository.ts
@@ -1,5 +1,6 @@
 import { CreateGuestbookDTO } from '../dtos/CreateGuestbookDTO'
 import { Guestbook } from '../models/Guestbook'
+import { GuestbookResponseSubset } from '../models/GuestbookResponse'
 
 export interface IGuestbooksRepository {
   createGuestbook(
@@ -7,7 +8,20 @@ export interface IGuestbooksRepository {
     guestbook: CreateGuestbookDTO
   ): Promise<number>
   getGuestbook(guestbookId: number): Promise<Guestbook>
-  getGuestbooksByCollectionId(collectionIdOrAlias: number | string): Promise<Guestbook[]>
+  getGuestbooksByCollectionId(
+    collectionIdOrAlias: number | string,
+    includeStats?: boolean,
+    includeInherited?: boolean
+  ): Promise<Guestbook[]>
+  getGuestbookResponsesByGuestbookId(
+    guestbookId: number,
+    limit?: number,
+    offset?: number
+  ): Promise<GuestbookResponseSubset>
+  downloadGuestbookResponsesByCollectionId(
+    collectionIdOrAlias: number | string,
+    guestbookId?: number
+  ): Promise<string>
   setGuestbookEnabled(
     collectionIdOrAlias: number | string,
     guestbookId: number,