+++ src/guestbooks/infra/repositories/GuestbooksRepository.ts
@@ -1,11 +1,14 @@
 import { ApiRepository } from '../../../core/infra/repositories/ApiRepository'
 import { CreateGuestbookDTO } from '../../domain/dtos/CreateGuestbookDTO'
+import { GuestbookResponsesDTO } from '../../domain/dtos/GuestbookResponsesDTO'
 import { Guestbook } from '../../domain/models/Guestbook'
+import { GuestbookResponseSubset } from '../../domain/models/GuestbookResponse'
 import { IGuestbooksRepository } from '../../domain/repositories/IGuestbooksRepository'
 
 export class GuestbooksRepository extends ApiRepository implements IGuestbooksRepository {
   private readonly guestbooksResourceName: string = 'guestbooks'
   private readonly datasetsResourceName: string = 'datasets'
+  private readonly dataversesResourceName: string = 'dataverses'
 
   public async createGuestbook(
     collectionIdOrAlias: number | string,
@@ -33,13 +36,60 @@ export class GuestbooksRepository extends ApiRepository implements IGuestbooksRe
   }
 
   public async getGuestbooksByCollectionId(
-    collectionIdOrAlias: number | string
+    collectionIdOrAlias: number | string,
+    includeStats = false,
+    includeInherited = false
   ): Promise<Guestbook[]> {
+    const queryParams = {
+      ...(includeStats ? { includeStats } : {}),
+      ...(includeInherited ? { includeInherited } : {})
+    }
+
     return this.doGet(
       this.buildApiEndpoint(this.guestbooksResourceName, `${collectionIdOrAlias}/list`),
-      true
+      true,
+      queryParams
     )
       .then((response) => response.data.data as Guestbook[])
+      .catch((error) => {
+        throw error
+      })
+  }
+
+  public async getGuestbookResponsesByGuestbookId(
+    guestbookId: number,
+    limit = 10,
+    offset = 0
+  ): Promise<GuestbookResponseSubset> {
+    return this.doGet(
+      this.buildApiEndpoint(this.guestbooksResourceName, 'responses', guestbookId),
+      true,
+      { limit, offset }
+    )
+      .then((response) => {
+        const responseData = response.data.data as GuestbookResponsesDTO
+
+        return {
+          guestbookResponses: responseData.responses,
+          totalGuestbookResponseCount: responseData.pagination?.totalResponses ?? 0
+        }
+      })
+      .catch((error) => {
+        throw error
+      })
+  }
+
+  public async downloadGuestbookResponsesByCollectionId(
+    collectionIdOrAlias: number | string,
+    guestbookId?: number
+  ): Promise<string> {
+    const endpoint = this.buildApiEndpoint(
+      this.dataversesResourceName,
+      `${collectionIdOrAlias}/guestbookResponses`
+    )
+
+    return this.doGet(endpoint, true, guestbookId ? { guestbookId } : {})
+      .then((response) => response.data as string)
       .catch((error) => {
         throw error
       })